Move mathematical types from std_types
