+++ src/util/std_types.h
@@ -56,42 +56,6 @@ class void_typet:public empty_typet
 {
 };
 
-/// Unbounded, signed integers (mathematical integers, not bitvectors)
-class integer_typet:public typet
-{
-public:
-  integer_typet():typet(ID_integer)
-  {
-  }
-};
-
-/// Natural numbers including zero (mathematical integers, not bitvectors)
-class natural_typet:public typet
-{
-public:
-  natural_typet():typet(ID_natural)
-  {
-  }
-};
-
-/// Unbounded, signed rational numbers
-class rational_typet:public typet
-{
-public:
-  rational_typet():typet(ID_rational)
-  {
-  }
-};
-
-/// Unbounded, signed real numbers
-class real_typet:public typet
-{
-public:
-  real_typet():typet(ID_real)
-  {
-  }
-};
-
 /// A reference into the symbol table
 class symbol_typet:public typet
 {
@@ -1821,113 +1785,6 @@ inline complex_typet &to_complex_type(typet &type)
   PRECONDITION(can_cast_type<complex_typet>(type));
   return static_cast<complex_typet &>(type);
 }
-
-/// A type for mathematical functions (do not confuse with functions/methods
-/// in code)
-class mathematical_function_typet:public typet
-{
-public:
-  // the domain of the function is composed of zero, one, or
-  // many variables
-  class variablet:public irept
-  {
-  public:
-    // the identifier is optional
-    irep_idt get_identifier() const
-    {
-      return get(ID_identifier);
-    }
-
-    void set_identifier(const irep_idt &identifier)
-    {
-      return set(ID_identifier, identifier);
-    }
-
-    typet &type()
-    {
-      return static_cast<typet &>(add(ID_type));
-    }
-
-    const typet &type() const
-    {
-      return static_cast<const typet &>(find(ID_type));
-    }
-  };
-
-  using domaint=std::vector<variablet>;
-
-  mathematical_function_typet(const domaint &_domain, const typet &_codomain)
-    : typet(ID_mathematical_function)
-  {
-    subtypes().resize(2);
-    domain() = _domain;
-    codomain() = _codomain;
-  }
-
-  domaint &domain()
-  {
-    return (domaint &)subtypes()[0].subtypes();
-  }
-
-  const domaint &domain() const
-  {
-    return (const domaint &)subtypes()[0].subtypes();
-  }
-
-  variablet &add_variable()
-  {
-    auto &d=domain();
-    d.push_back(variablet());
-    return d.back();
-  }
-
-  /// Return the codomain, i.e., the set of values that the function maps to
-  /// (the "target").
-  typet &codomain()
-  {
-    return subtypes()[1];
-  }
-
-  /// \copydoc codomain()
-  const typet &codomain() const
-  {
-    return subtypes()[1];
-  }
-};
-
-/// Check whether a reference to a typet is a \ref mathematical_function_typet.
-/// \param type Source type.
-/// \return True if \param type is a \ref mathematical_function_typet.
-template <>
-inline bool can_cast_type<mathematical_function_typet>(const typet &type)
-{
-  return type.id() == ID_mathematical_function;
-}
-
-/// \brief Cast a typet to a \ref mathematical_function_typet
-///
-/// This is an unchecked conversion. \a type must be known to be \ref
-/// mathematical_function_typet. Will fail with a precondition violation if type
-/// doesn't match.
-///
-/// \param type Source type.
-/// \return Object of type \ref mathematical_function_typet.
-inline const mathematical_function_typet &
-  to_mathematical_function_type(const typet &type)
-{
-  PRECONDITION(can_cast_type<mathematical_function_typet>(type));
-  return static_cast<const mathematical_function_typet &>(type);
-}
-
-/// \copydoc to_mathematical_function_type(const typet &)
-inline mathematical_function_typet &
-  to_mathematical_function_type(typet &type)
-{
-  PRECONDITION(can_cast_type<mathematical_function_typet>(type));
-  return static_cast<mathematical_function_typet &>(type);
-}
-
-bool is_number(const typet &type);
 
 bool is_constant_or_has_constant_components(
   const typet &type,
